Добавил фронт и контроллер для генерации нового кода, зачисление пока что не работает

+++ backend/Shop/Interfaces/ITopUpService.cs
@@ -4,6 +4,6 @@
     {
         Task<bool> ApplyTopUpCodeAsync(string code, int userId, CancellationToken cancellationToken);
 
-        Task<string> CreateTopUpCodeAsync(int amountValue, CancellationToken cancellationToken);
+        Task<string> CreateTopUpCodeAsync(decimal amountValue, CancellationToken cancellationToken);
     }
 }
